refactor(consensus): execute engine operations directly
Replace the engine task queue with direct serialized engine operations for build, insert, consolidate, finalize, and forkchoice handling.

Carry through the existing service request, tracing, checkpoint, and test coverage updates in one rebased commit.
