+++ crates/consensus/engine/src/operations/mod.rs
@@ -1,0 +1,33 @@
+//! Direct execution-layer operations.
+
+mod engine;
+pub use engine::Engine;
+
+mod severity;
+pub use severity::{EngineTaskError, EngineTaskErrorSeverity};
+
+mod retry;
+
+mod build;
+pub use build::{BuildTaskError, EngineBuildError};
+
+mod payload;
+pub use payload::SealTaskError;
+
+mod insert;
+pub use insert::{InsertPayloadSafety, InsertTaskError, InsertTaskResult};
+
+mod consolidate;
+pub use consolidate::{ConsolidateInput, ConsolidateTaskError};
+
+mod delegated_forkchoice;
+pub use delegated_forkchoice::{DelegatedForkchoiceTaskError, DelegatedForkchoiceUpdate};
+
+mod finalize;
+pub use finalize::FinalizeTaskError;
+
+mod reset;
+pub use reset::EngineResetError;
+
+mod synchronize;
+pub use synchronize::{SynchronizeTask, SynchronizeTaskError};
