--- crates/consensus/engine/src/task_queue/mod.rs
@@ -1,7 +1,0 @@
-//! The [`Engine`] task queue and the [`EngineTask`]s it can execute.
-
-mod core;
-pub use core::{Engine, EngineResetError};
-
-mod tasks;
-pub use tasks::*;
