--- crates/consensus/engine/src/task_queue/tasks/consolidate/mod.rs
@@ -1,10 +1,0 @@
-//! Consolidation Task
-
-mod error;
-pub use error::ConsolidateTaskError;
-
-mod task;
-pub use task::{ConsolidateInput, ConsolidateTask};
-
-#[cfg(test)]
-mod task_test;
