--- crates/consensus/engine/src/task_queue/tasks/insert/error.rs
@@ -1,51 +1,0 @@
-//! Contains the error types for the [`InsertTask`].
-//!
-//! [InsertTask]: crate::InsertTask
-
-use alloy_rpc_types_engine::PayloadStatusEnum;
-use alloy_transport::{RpcError, TransportErrorKind};
-use base_common_rpc_types_engine::BasePayloadError;
-use base_protocol::FromBlockError;
-
-use crate::{
-    EngineTaskError, SynchronizeTaskError, task_queue::tasks::task::EngineTaskErrorSeverity,
-};
-
-/// An error that occurs when running the [`InsertTask`].
-///
-/// [InsertTask]: crate::InsertTask
-#[derive(Debug, thiserror::Error)]
-pub enum InsertTaskError {
-    /// Error converting a payload into a block.
-    #[error(transparent)]
-    FromBlockError(#[from] BasePayloadError),
-    /// Failed to insert new payload.
-    #[error("Failed to insert new payload: {0}")]
-    InsertFailed(RpcError<TransportErrorKind>),
-    /// Unexpected payload status
-    #[error("Unexpected payload status: {0}")]
-    UnexpectedPayloadStatus(PayloadStatusEnum),
-    /// Error converting the payload + chain genesis into an L2 block info.
-    #[error(transparent)]
-    L2BlockInfoConstruction(#[from] FromBlockError),
-    /// The forkchoice update call to consolidate the block into the engine state failed.
-    #[error(transparent)]
-    ForkchoiceUpdateFailed(#[from] SynchronizeTaskError),
-    /// The forkchoice update completed without advancing the unsafe head to the inserted payload.
-    #[error("Forkchoice update did not advance to the inserted payload")]
-    ForkchoiceUpdateDidNotAdvance,
-}
-
-impl EngineTaskError for InsertTaskError {
-    fn severity(&self) -> EngineTaskErrorSeverity {
-        match self {
-            Self::FromBlockError(_) | Self::L2BlockInfoConstruction(_) => {
-                EngineTaskErrorSeverity::Critical
-            }
-            Self::InsertFailed(_)
-            | Self::UnexpectedPayloadStatus(_)
-            | Self::ForkchoiceUpdateDidNotAdvance => EngineTaskErrorSeverity::Temporary,
-            Self::ForkchoiceUpdateFailed(inner) => inner.severity(),
-        }
-    }
-}
