--- crates/consensus/engine/src/task_queue/tasks/insert/mod.rs
@@ -1,7 +1,0 @@
-//! Task to insert a payload into the execution engine.
-
-mod task;
-pub use task::{InsertPayloadSafety, InsertTask, InsertTaskResult};
-
-mod error;
-pub use error::InsertTaskError;
