--- crates/consensus/engine/src/task_queue/tasks/seal/mod.rs
@@ -1,7 +1,0 @@
-//! Task and its associated types for importing a block that has been started.
-
-mod task;
-pub use task::SealTask;
-
-mod error;
-pub use error::SealTaskError;
