--- crates/consensus/engine/src/task_queue/tasks/synchronize/error.rs
@@ -1,36 +1,0 @@
-//! Contains error types for the [`crate::SynchronizeTask`].
-
-use alloy_rpc_types_engine::PayloadStatusEnum;
-use alloy_transport::{RpcError, TransportErrorKind};
-use thiserror::Error;
-
-use crate::{EngineTaskError, task_queue::tasks::task::EngineTaskErrorSeverity};
-
-/// An error that occurs when running the [`crate::SynchronizeTask`].
-#[derive(Debug, Error)]
-pub enum SynchronizeTaskError {
-    /// The forkchoice update call to the engine api failed.
-    #[error("Forkchoice update engine api call failed due to an RPC error: {0}")]
-    ForkchoiceUpdateFailed(RpcError<TransportErrorKind>),
-    /// The finalized head is behind the unsafe head.
-    #[error("Invalid forkchoice state: unsafe head {0} is ahead of finalized head {1}")]
-    FinalizedAheadOfUnsafe(u64, u64),
-    /// The forkchoice state is invalid.
-    #[error("Invalid forkchoice state")]
-    InvalidForkchoiceState,
-    /// The payload status is unexpected.
-    #[error("Unexpected payload status: {0}")]
-    UnexpectedPayloadStatus(PayloadStatusEnum),
-}
-
-impl EngineTaskError for SynchronizeTaskError {
-    fn severity(&self) -> EngineTaskErrorSeverity {
-        match self {
-            Self::FinalizedAheadOfUnsafe(_, _) => EngineTaskErrorSeverity::Critical,
-            Self::ForkchoiceUpdateFailed(_) | Self::UnexpectedPayloadStatus(_) => {
-                EngineTaskErrorSeverity::Temporary
-            }
-            Self::InvalidForkchoiceState => EngineTaskErrorSeverity::Reset,
-        }
-    }
-}
