+++ crates/consensus/service/src/actors/engine/request.rs
@@ -1,7 +1,8 @@
 use alloy_rpc_types_engine::PayloadId;
 use base_common_rpc_types_engine::BaseExecutionPayloadEnvelope;
 use base_consensus_engine::{
-    BuildTaskError, ConsolidateInput, EngineQueries, InsertTaskError, SealTaskError,
+    BuildTaskError, ConsolidateInput, DelegatedForkchoiceUpdate, EngineQueries, InsertTaskError,
+    SealTaskError,
 };
 use base_protocol::{AttributesWithParent, L2BlockInfo};
 use opentelemetry::Context;
@@ -54,6 +55,8 @@ pub enum EngineActorRequest {
     /// Request to consolidate using a safe L2 signal from attributes or delegated safe-block
     /// derivation
     ProcessSafeL2SignalRequest(ConsolidateInput),
+    /// Request to apply delegated follow-node safe/finalized labels together.
+    ProcessDelegatedForkchoiceUpdateRequest(Box<DelegatedForkchoiceUpdate>),
     /// Request to finalize the L2 block at the provided block number.
     ProcessFinalizedL2BlockNumberRequest(Box<u64>),
     /// Request to insert the provided external unsafe block.
