+++ crates/consensus/service/src/actors/mod.rs
@@ -16,9 +16,10 @@ mod engine;
 pub use engine::MockEngineDerivationClient;
 pub use engine::{
     BootstrapRole, BuildRequest, EngineActor, EngineActorRequest, EngineClientError,
-    EngineClientResult, EngineConfig, EngineDerivationClient, EngineError, EngineProcessor,
-    EngineProcessorOptions, EngineRequestReceiver, EngineRpcProcessor, EngineRpcRequest,
-    GetPayloadRequest, InsertUnsafePayloadRequest, QueuedEngineDerivationClient, ResetRequest,
+    EngineClientResult, EngineConfig, EngineDerivationClient, EngineError, EngineOperationError,
+    EngineProcessor, EngineProcessorOptions, EngineRequestReceiver, EngineRpcProcessor,
+    EngineRpcRequest, GetPayloadRequest, InsertUnsafePayloadRequest, QueuedEngineDerivationClient,
+    ResetRequest,
 };
 
 mod rpc;
