+++ crates/consensus/service/src/service/node.rs
@@ -244,8 +244,7 @@ impl RollupNode {
     {
         let engine_state = EngineState::default();
         let (engine_state_tx, engine_state_rx) = watch::channel(engine_state);
-        let (engine_queue_length_tx, engine_queue_length_rx) = watch::channel(0);
-        let engine = Engine::new(engine_state, engine_state_tx, engine_queue_length_tx);
+        let engine = Engine::new(engine_state, engine_state_tx);
 
         let mode = self.mode();
         let checkpoint_reader: Arc<dyn ForkchoiceCheckpointReader> =
@@ -270,7 +269,6 @@ impl RollupNode {
             Arc::clone(&engine_client),
             Arc::clone(&self.config),
             engine_state_rx,
-            engine_queue_length_rx,
         );
 
         let engine_actor =
